requirements.txt
certifi==2021.10.8
click==8.0.3
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.21.5
pandas==1.3.5
patsy==0.5.2
plotly==5.5.0
plotly-express==0.4.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2021.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
SQLAlchemy==1.3.23
statsmodels==0.13.1
tenacity==8.0.1
threadpoolctl==3.0.0
typing_extensions==4.0.1
Werkzeug==2.0.2
zipp==3.7.0
